complate history page

+++ src/main/java/com/example/Custom/domain/OrderDetail.java
@@ -17,7 +17,7 @@ public class OrderDetail {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private long quantity;
+    private int quantity;
     private double price;
     @ManyToOne
     @JoinColumn(name = "order_id")
